pyproject.toml
[tool.poetry]
name = "data-agent"
version = "0.1.0"
description = "A RAG-powered AI agent specialized in Data Engineering"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "data_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.0"
llama-index = "^0.9.0"
pydantic = "^2.0.0"
openai = "^1.0.0"
chromadb = "^0.4.0"
unstructured = "^0.10.0"
nltk = "^3.8.0"
transformers = "^4.35.0"
sentence-transformers = "^2.2.0"
rich = "^13.0.0"
typer = "^0.9.0"
python-dotenv = "^1.0.0"
tqdm = "^4.66.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"
mypy = "^1.0.0"

[tool.poetry.scripts]
data-agent = "data_agent.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
